requirements.txt
torch==2.4.1
torchvision==0.19.1
timm==1.0.11
albumentations==1.4.18
opencv-python-headless==4.10.0.84
iterative-stratification==0.1.9
monai==1.3.2
monai-generative==0.2.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pyyaml==6.0.2
torchmetrics==1.5.0
pytorch-fid==0.3.0
matplotlib==3.9.2
pyarrow==17.0.0
tqdm==4.66.5
